combine dungeon_movement
