+++ src/code_8077274_1.c
@@ -5,7 +5,7 @@
 #include "constants/type.h"
 #include "dungeon_entity.h"
 #include "dungeon_global_data.h"
-#include "dungeon_movement_1.h"
+#include "dungeon_movement.h"
 #include "dungeon_pokemon_attributes.h"
 #include "dungeon_random.h"
 #include "dungeon_items.h"
