+++ src/dungeon_engine.c
@@ -5,7 +5,7 @@
 #include "dungeon_global_data.h"
 #include "dungeon_ai.h"
 #include "dungeon_leader.h"
-#include "dungeon_movement_1.h"
+#include "dungeon_movement.h"
 #include "dungeon_util.h"
 
 extern s16 gSpeedTurns[2][25];
